requirements.txt
torch>=2.0.0
transformers>=4.30.0
pillow>=9.0.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
numpy>=1.21.0
huggingface-hub>=0.15.0
safetensors>=0.3.0